Renamed 'adjacency.py' to 'celllist.py'
