+++ src/biotite/structure/__init__.py
@@ -80,9 +80,9 @@ visualisation.
 
 __author__ = "Patrick Kunzmann"
 
-from .adjacency import *
 from .atoms import *
 from .bonds import *
+from .celllist import *
 from .compare import *
 from .error import *
 from .filter import *
