+++ src/biotite/structure/__init__.pyi
@@ -5,6 +5,7 @@
 from .adjacency import *
 from .atoms import *
 from .bonds import *
+from .celllist import *
 from .compare import *
 from .error import *
 from .filter import *
